requirements.txt
# Зависимости для проекта
PyYAML>=6.0